+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added `nodoc!` method to `ActionDefinition`, `ResourceDefinition` to hide actions and resources from the generated documentation.
 * Added descriptions to the default HTTP responses.
 * Replaced Ruport dependency in `praxis:routes` rake task with TerminalTable.
+* Fixed doc browser issue when attributes defaulting to false wouldn't display the default section.
 
 ## 0.11.2