+++ CHANGELOG.md
@@ -16,6 +16,8 @@
   * Configured with boolean `praxis.x_cascade` that defaults to true.
   * When enabled, Praxis will add an 'X-Cascade: pass' header to the response when the request was not routable to an action. It is not added if the action explicitly returns a `NotFound` response.
 * Fixed bug in request handling where `after` callbacks were being executed, even if the stage returned a response.
+* Added a handy option to tie an action route to match any HTTP verb.
+  * Simply use `any` as the verb when you define it (i.e. any '/things/:id' )
 
 ## 0.11.2